fix: Rename get_generators to operator_generator everywhere
